test(proxy): run managed passthrough limit tests in CI
